This is a first pass at introducing the concept of readonly_vars
Adding this member to the struct may aid in correctly handling read-only
variables in #1205.

I don't love that we need to be sure to initialize this member
everywhere a var is initialized.

May need to make this function public to be useful to #1205.

I don't consider `default_transaction_read_only` to be read-only
because clients CAN set it but it can also be set by the server and so
maybe that's a bit special. I think users of this feature would benefit
from #1205 updating vars as the server connection vars change.